app/ai_functions/requirements.txt
openai 
pillow 
requests 
gTTS 
moviepy 
keybert
boto3
fastapi
uvicorn
colorama
srt